+++ src/azure-cli/azure/cli/command_modules/appservice/custom.py
@@ -2301,16 +2301,16 @@ def validate_range_of_int_flag(flag_name, value, min_val, max_val):
 
 
 def create_function(cmd, resource_group_name, name, storage_account, plan=None,
-                    os_type=None, version=None, runtime=None, runtime_version=None, consumption_plan_location=None,
+                    os_type=None, functions_version=None, runtime=None, runtime_version=None, consumption_plan_location=None,
                     app_insights=None, app_insights_key=None, disable_app_insights=None, deployment_source_url=None,
                     deployment_source_branch='master', deployment_local_git=None,
                     docker_registry_server_password=None, docker_registry_server_user=None,
                     deployment_container_image_name=None, tags=None):
     # pylint: disable=too-many-statements, too-many-branches
-    if version is None:
-        logger.warning("No app version specified so defaulting to 2. In the future, specifying a version will "
-                       "be required. To create a 2.x function you would pass in the flag `--version 2`")
-        version = '2'
+    if functions_version is None:
+        logger.warning("No functions version specified so defaulting to 2. In the future, specifying a version will "
+                       "be required. To create a 2.x function you would pass in the flag `--functions_version 2`")
+        functions_version = '2'
     if deployment_source_url and deployment_local_git:
         raise CLIError('usage error: --deployment-source-url <url> | --deployment-local-git')
     if bool(plan) == bool(consumption_plan_location):
@@ -2364,11 +2364,11 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
     if runtime_version is not None:
         if runtime is None:
             raise CLIError('Must specify --runtime to use --runtime-version')
-        allowed_versions = RUNTIME_TO_IMAGE_FUNCTIONAPP[version][runtime].keys()
+        allowed_versions = RUNTIME_TO_IMAGE_FUNCTIONAPP[functions_version][runtime].keys()
         if runtime_version not in allowed_versions:
-            raise CLIError('--runtime-version {} is not supported for the selected --runtime {}. '
-                           'Supported versions are: {}'
-                           .format(runtime_version, runtime, ', '.join(allowed_versions)))
+            raise CLIError('--runtime-version {} is not supported for the selected --runtime {} and '
+                           '--functions_version {}. Supported versions are: {}'
+                           .format(runtime_version, runtime, functions_version, ', '.join(allowed_versions)))
 
     con_string = _validate_and_get_connection_string(cmd.cli_ctx, resource_group_name, storage_account)
 
@@ -2390,19 +2390,19 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
             else:
                 site_config.app_settings.append(NameValuePair(name='WEBSITES_ENABLE_APP_SERVICE_STORAGE',
                                                               value='true'))
-                if runtime not in RUNTIME_TO_IMAGE_FUNCTIONAPP[version].keys():
+                if runtime not in RUNTIME_TO_IMAGE_FUNCTIONAPP[functions_version].keys():
                     raise CLIError("An appropriate linux image for runtime:'{}' was not found".format(runtime))
         if deployment_container_image_name is None:
-            site_config.linux_fx_version = _get_linux_fx_functionapp(is_consumption, version, runtime, runtime_version)
+            site_config.linux_fx_version = _get_linux_fx_functionapp(is_consumption, functions_version, runtime, runtime_version)
     else:
         functionapp_def.kind = 'functionapp'
     # adding appsetting to site to make it a function
     site_config.app_settings.append(NameValuePair(name='FUNCTIONS_EXTENSION_VERSION',
-                                                  value=_get_runtime_version_functionapp(version)))
+                                                  value=_get_functions_extension_version_functionapp(functions_version)))
     site_config.app_settings.append(NameValuePair(name='AzureWebJobsStorage', value=con_string))
     site_config.app_settings.append(NameValuePair(name='AzureWebJobsDashboard', value=con_string))
     site_config.app_settings.append(NameValuePair(name='WEBSITE_NODE_DEFAULT_VERSION',
-                                                  value=_get_website_node_version_functionapp(version,
+                                                  value=_get_website_node_version_functionapp(functions_version,
                                                                                               runtime,
                                                                                               runtime_version)))
 
@@ -2455,27 +2455,27 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
     return functionapp
 
 
-def _get_runtime_version_functionapp(version):
-    if version is not None:
-        return '~{}'.format(version)
+def _get_functions_extension_version_functionapp(functions_version):
+    if functions_version is not None:
+        return '~{}'.format(functions_version)
     return '~2'
 
 
-def _get_linux_fx_functionapp(is_consumption, version, runtime, runtime_version):
+def _get_linux_fx_functionapp(is_consumption, functions_version, runtime, runtime_version):
     if runtime_version is None:
-        runtime_version = RUNTIME_TO_DEFAULT_VERSION_FUNCTIONAPP[version][runtime]
+        runtime_version = RUNTIME_TO_DEFAULT_VERSION_FUNCTIONAPP[functions_version][runtime]
     if is_consumption:
         return '{}|{}'.format(runtime.upper(), runtime_version)
     # App service or Elastic Premium
-    return _format_fx_version(RUNTIME_TO_IMAGE_FUNCTIONAPP[version][runtime][runtime_version])
+    return _format_fx_version(RUNTIME_TO_IMAGE_FUNCTIONAPP[functions_version][runtime][runtime_version])
 
 
-def _get_website_node_version_functionapp(version, runtime, runtime_version):
+def _get_website_node_version_functionapp(functions_version, runtime, runtime_version):
     if runtime is None or runtime != 'node':
-        return NODE_VERSION_DEFAULT_FUNCTIONAPP[version]
+        return NODE_VERSION_DEFAULT_FUNCTIONAPP[functions_version]
     if runtime_version is not None:
         return '~{}'.format(runtime_version)
-    return NODE_VERSION_DEFAULT_FUNCTIONAPP[version]
+    return NODE_VERSION_DEFAULT_FUNCTIONAPP[functions_version]
 
 
 def try_create_application_insights(cmd, functionapp):